iproute2: tc: reduce size of dynamic symbol table
In the case of SHARED_LIBS=y, don't use -export-dynamic to place *all*
symbols into the dynamic symbol table. Instead, use --dynamic-list to
export a smaller set of symbols similar to that defined in static-syms.h
for the case of SHARED_LIBS=n.

This avoids package size increases of up to 40 KB in some cases.

Signed-off-by: Tony Ambardar <itugrok@yahoo.com>