squash! kernel/modules: add missing dependency to kmod-sched-core package
Add the matchall kernel module cls_matchall introduced in kernel 4.8.

The matchall classifier matches every packet and allows the user to apply
actions on it. It is a simpler, more efficient replacement for the common
but cryptic tc classifier idiom "u32 match u32 0 0".

This filter is very useful in cases where every packet should be matched.
For example, packet mirroring (SPAN) can be setup very easily
using that filter.