bpftool: add utility package for Linux eBPF subsystem
bpftool is a userspace tool included with the kernel sources, and widely
used for introspection and manipulation of eBPF programs and maps.

This package targets the 4.19.x and 5.4.x stable kernels. Two variants are
built: a 'full' version which supports object disassembly and depends on
libbfd/libopcodes (~500KB); and a 'minimal' version without disassembly
functions and dependencies. The default 'minimal' variant is otherwise
fully functional.

Signed-off-by: Tony Ambardar <itugrok@yahoo.com>